+++ endpoints/index.js
@@ -1,0 +1,3 @@
+const users = require('./users')
+
+module.exports = { users }
